pyproject.toml
[project]
name = "group-file-admin"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "zhongwen-4", email = "2401128923@qq.com"},
]
dependencies = [
    "nonebot-plugin-localstore>=0.7.3",
    "nonebot-plugin-waiter>=0.8.1",
    "httpx>=0.28.1",
    "nonebot2>=2.4.1",
    "nonebot-adapter-onebot>=2.4.6",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
